refactor(backend): FASE 4 - migrate cache & messaging imports to modern paths
🎯 Complete migration of 47 files from deprecated service stubs to domain-driven architecture

## Changes Summary

### Unified Cache Migration (6 active files)
- OLD: from app.utils.unified_cache import ...
- NEW: from app.infrastructure.cache import ...
- Updated: patient.py, auth.py, cache_middleware.py, performance.py, patients_flow.py, unified_cache.py

### MessageSender Migration (19 active files)
- OLD: from app.services.message_sender import MessageSender
- NEW: from app.domain.messaging.delivery import MessageSender
- Updated: All messaging, task, and flow-related modules

### MessageFactory Migration (8 active files)
- OLD: from app.services.message_factory import MessageFactory
- NEW: from app.domain.messaging.core import MessageFactory
- Updated: Quiz, AB testing, and service modules

## Archived Files (154 lines)
- app/services/message_sender.py → legacy/services_archive_2025-11-09/
- app/services/message_factory.py → legacy/services_archive_2025-11-09/
- app/utils/unified_cache.py → legacy/services_archive_2025-11-09/

## Additional Updates
- Fixed scripts/cleanup_legacy_cache.py (removed self-import)
- Updated documentation with new import paths
- Created migration mapping and completion report

## Verification
✅ All target modules exist and exports configured
✅ No remaining active imports of deprecated paths
✅ Import structure validated (47 files)
✅ Zero breaking changes

Total cleanup: FASES 1-4 complete (3,531 lines removed)

Refs: docs/FASE4_COMPLETION_REPORT.md, docs/FASE4_MIGRATION_MAPPING.md
