+++ backend-hormonia/app/domain/agents/quiz/notification_manager.py
@@ -12,7 +12,7 @@ from enum import Enum
 from sqlalchemy.orm import Session
 
 from app.models.message import Message, MessageType, MessageDirection, MessageStatus
-from app.services.message_sender import MessageSender
+from app.domain.messaging.delivery import MessageSender
 
 
 class QuizAdaptationType(Enum):
